video/video.ino: Enhance text viewport handling
To support video resolutions where the font does not exactly fit
(for example 400x300 resolution with 8 pixel high font) or other
font sizes that don't exactly fill the screen, adjust the full
screen viewport to take this into account.

Change vdu_resetViewports() to subtract any remainder from the
width and height of the viewport. For example, if the screen is
300 pixels high and the font is 8 pixels high then 300 % 8 will
be 4. Therefore, the viewport height will be 296 (0-395) rather
than 300.

Since we now use fontW and fontH in vdu_resetViewports(),
amend change_mode() not to call cls() until after fontW and
fontH are known.